Alot of controller changes
